feat: Complete StatisticsDashboard.js splitting (Task 6)
- Refactored StatisticsDashboard.js to lightweight main controller (2,596→1,428 words)
- Created DashboardWidgetRenderer.js with all widget implementations (1,279 words)
- Applied Main Controller Pattern with dependency injection
- Extracted 6 widget classes: KeyMetricsWidget, RecentAchievementsWidget,
  GrowthTrendsWidget, PlayStyleWidget, PerformanceChartWidget, StatisticsBreakdownWidget
- Modified createWidgets() to use widgetRenderer.createWidget()
- Maintained all public APIs through delegation
- Both components now under 2,500-word MCP compatibility limit

Task 6 Complete: StatisticsDashboard.js split into 2 components
✅ DashboardWidgetRenderer.js (1,279 words) - widget implementations
✅ StatisticsDashboard.js (1,428 words) - main controller & orchestration
